requirements.txt
playwright
python-dotenv
# langchain-openai # No longer needed 